% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_obs_eval.R
\name{collect_obs_eval}
\alias{collect_obs_eval}
\title{Get observational data for model evaluation}
\usage{
collect_obs_eval(
  siteinfo,
  settings,
  adf = NULL,
  mdf = NULL,
  ddf = NULL,
  hrdf = NULL,
  hhdf = NULL
)
}
\arguments{
\item{siteinfo}{A data frame containing site meta info (rows for sites). Required columns are: \code{"sitename", "date_start", "date_end", "lon", "lat", "elv"}.}

\item{settings}{A list specifying evaluation settings (see vignette eval_sofun.pdf for more information and examples)}

\item{adf}{A nested data frame with rows for sites and time series (annual) for each site nested inside the column \code{data}.}

\item{mdf}{A nested data frame with rows for sites and time series (monthly) for each site nested inside the column \code{data}.}

\item{ddf}{A nested data frame with rows for sites and time series (daily) for each site nested inside the column \code{data}.}

\item{hrdf}{A nested data frame with rows for sites and time series (hourly) for each site nested inside the column \code{data}.}

\item{hhdf}{A nested data frame with rows for sites and time series (half-hourly) for each site nested inside the column \code{data}.}
}
\value{
A list containing nested data frames of observed values aggregated to several temporal scales (ddf for daily, xdf for X-daily
(determined by argument \code{agg}), mdf for monthly, adf for annual).
}
\description{
Gets observational data for model evaluation, given the benchmarking variable, and data source used
for the evaluation. This information is specified in the evaluation settings (argument \code{settings}).
}
\examples{
\dontrun{
obs <- collect_obs_eval(settings = NULL, adf = NULL, mdf = NULL, ddf = NULL, hrdf = NULL, hhdf = NULL, agg = 8)
}

}
