% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_droughtresponse.R
\name{eval_droughtresponse}
\alias{eval_droughtresponse}
\title{Evaluate the model's drought response}
\usage{
eval_droughtresponse(
  df,
  df_flue,
  before,
  after,
  leng_threshold,
  nbins,
  do_norm = FALSE
)
}
\arguments{
\item{df}{A data frame containing all data for all sites, modelled and observed}

\item{obsvarnam}{The variable name of the observational data, given as a column in \code{df}}

\item{modvarnam}{The variable name of the modelled data, given as a column in \code{df}}
}
\value{
A data frame of data aligned by drought events and aggregated across events (for different quantiles)
}
\description{
xxx
}
\examples{
xxx

}
