% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sofun.R
\name{eval_sofun}
\alias{eval_sofun}
\title{Evaluates SOFUN model outputs.}
\usage{
eval_sofun(
  mod,
  settings,
  obs_eval = NA,
  overwrite = TRUE,
  doplot = FALSE,
  light = FALSE
)
}
\arguments{
\item{mod}{Object returned by \link{runread_pmodel_f}. This is a nested
dataframe with sites along rows and a nested column \code{"data"} containing
model outputs with columns \code{"date"} (date object created by
\code{lubridate::ymd()}) and \code{"varnam"}.
where \code{"varnam"} corresponds to \code{names(settings$benchmark)}.}

\item{settings}{A list specifying evaluation settings
(see vignette eval_sofun.pdf for more information and examples)}

\item{obs_eval}{(Optional) A named list of data frames containing
observational data for each sites. The names of list elements corresponds
to site names. Defaults to \code{NA}}

\item{overwrite}{(Optional) A logical specifying whether temporary data
stored in \code{./tmpdir} should be overwritten. Defaults to \code{TRUE}.}

\item{doplot}{(Optional) A logical specifying whether plots should be saved.
Defaults to \code{FALSE}.}

\item{light}{(Optional) A logical specifying whether reduced data should
saved. Defaults to \code{FALSE}.}
}
\value{
A list containing data frames of modelled and observed values
 aggregated to several temporal scales (ddf for daily, xdf for X-daily, mdf
 for monthly, adf for annual), data frames of respective performance metrics,
 and functions for plotting respective data.
}
\description{
Calculates a set of perfomance metrics for model outputs, compared against
observational data. Currently only evaluations of GPP model outputs, compared
agains FLUXNET 2015 data, are implemented.
}
