% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_days.R
\name{filter_days}
\alias{filter_days}
\title{Filter daily observations}
\usage{
filter_days(ddf, filter_days_source, path_gepisat)
}
\arguments{
\item{ddf}{xxx}

\item{filter_days_source}{xxx}

\item{path_gepisat}{xxx}
}
\value{
xxx
}
\description{
Filters observational data points for \code{gpp_obs} (hard-coded) in one data source, based on data availability from another data source.
}
\examples{
xxx
}
