% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_drivers.R
\name{format_drivers}
\alias{format_drivers}
\title{Format and run p-model drivers on Euler}
\usage{
format_drivers(
  siteinfo,
  params_siml = list(spinup = TRUE, spinupyears = 10, recycle = 1, soilmstress = FALSE,
    tempstress = FALSE, calc_aet_fapar_vpd = FALSE, in_ppfd = TRUE, in_netrad = FALSE,
    outdt = 1, ltre = FALSE, ltne = FALSE, ltrd = FALSE, ltnd = FALSE, lgr3 = TRUE, lgn3
    = FALSE, lgr4 = FALSE),
  params_modl = list(kphio = 0.09423773, soilm_par_a = 0.33349283, soilm_par_b =
    1.45602286, tau_acclim_tempstress = 10, par_shape_tempstress = 0),
  df_soiltexture = bind_rows(top = tibble(layer = "top", fsand = 0.4, fclay = 0.3, forg
    = 0.1, fgravel = 0.1), bottom = tibble(layer = "bottom", fsand = 0.4, fclay = 0.3,
    forg = 0.1, fgravel = 0.1)),
  bias_correction = FALSE,
  verbose = TRUE,
  run_model = FALSE,
  flux_data
)
}
\arguments{
\item{siteinfo}{data frame using minimum information required
being five columns: sitename, lon, lat, start_year, end_year}

\item{params_siml}{simulation parameters (preset)}

\item{params_modl}{model parameters (preset)}

\item{df_soiltexture}{soil data specifics (preset)}

\item{bias_correction}{bolean TRUE/FALSE, when using high resolution
locations set the bias correction to TRUE, for global runs leave FALSE
(default)}

\item{verbose}{provide verbose output (default = FALSE)}

\item{run_model}{shoud model evaluation be run, which returns evaluated
data instead of driver files?}
}
\value{
returns an rsofun compatible driver file for the provided
 sites
}
\description{
Takes site information for a single
or multiple sites and grabs all data
required for a p-model run in rsofun.
}
\details{
Parameter settings are provided as
arguments, but could be altered after
the fact if desired.
}
