% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get statistics}
\usage{
get_stats(mod, obs)
}
\arguments{
\item{mod}{A numeric vector of modelled values}

\item{obs}{A numeric vector of observed values}
}
\value{
A single row of a data frame (tibble) with columns corresponding to statistics
}
\description{
Returns multiple model performance statistics
}
\examples{
stats <- get_stats(mod, obs)
}
