% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_calib_eval_sofun.R
\name{sbs_calib_eval_sofun_bysite}
\alias{sbs_calib_eval_sofun_bysite}
\title{Conducts an site-by-site calibration (helper function)}
\usage{
sbs_calib_eval_sofun_bysite(
  evalsite,
  settings_calib,
  settings_eval,
  settings_sims,
  settings_input,
  df_drivers,
  ddf_obs_calib,
  obs_eval,
  overwrite
)
}
\arguments{
\item{evalsite}{which site to evaluate}

\item{settings_calib}{A list containing model calibration settings.
See vignette_rsofun.pdf for more information and examples.}

\item{settings_eval}{A list specifying evaluation settings
(see vignette eval_sofun.pdf for more information and examples)}

\item{settings_sims}{A list containing model simulation settings from
\code{\link{prepare_setup_sofun}}.  See vignette_rsofun.pdf for more
information and examples.}

\item{settings_input}{additional settings}

\item{df_drivers}{driver data}

\item{ddf_obs_calib}{A data frame containing observational data used for
model calibration. Created by function \code{get_obs_calib()}}

\item{obs_eval}{A data frame containing observational data used for model
evaluation Created by function \code{get_obs_eval()}}

\item{overwrite}{A boolean, defaults to \code{FALSE}.}
}
\value{

}
\description{
function to assist in the site by site comparison
}
